moved createMarkup into utils class
