+++ lib/__tests__/utils.test.js
@@ -1,72 +1,50 @@
 import assert from 'assert';
 import * as utils from '../utils';
 
-import { isPropertyUnitfull } from '../properties';
 
-const hyphenate = (stringToBeHyphenated) => {
-  return stringToBeHyphenated
-    .replace( /([a-z])([A-Z])/g, '$1-$2' )
-    .toLowerCase();
-};
 
-const isNumber = (value) => !isNaN(value);
-
-const createMarkup = (styleDeclaration) => {
-  const rules = Object.keys(styleDeclaration).map((cssProperty) => {
-    const rawCssValue = styleDeclaration[cssProperty];
-    const hyphenatedProperty = hyphenate(cssProperty);
-
-    let cssValue = rawCssValue;
-    if(isNumber(cssValue) && isPropertyUnitfull(hyphenatedProperty)) {
-      cssValue += 'px';
-    }
-
-    return `${hyphenatedProperty}:${cssValue};`;
-  });
-  return rules.join('');
-};
 
 describe('utils', () => {
   describe('#hyphenateStyleName', () => {
     it('hyphanates fontSize', () => {
-      const hyphenatedStyleName = hyphenate('fontSize');
+      const hyphenatedStyleName = utils.hyphenate('fontSize');
       assert.equal(hyphenatedStyleName, 'font-size');
     });
 
     it('doesn\'t hyphenates opacity', () => {
-      const hyphenatedStyleName = hyphenate('opacity');
+      const hyphenatedStyleName = utils.hyphenate('opacity');
       assert.equal(hyphenatedStyleName, 'opacity');
     });
 
     it('doesn\'t hyphenates already hyphanated property', () => {
-      const hyphenatedStyleName = hyphenate('font-size');
+      const hyphenatedStyleName = utils.hyphenate('font-size');
       assert.equal(hyphenatedStyleName, 'font-size');
     });
 
     it('supports vendor prefixes', () => {
-      const hyphenatedStyleName = hyphenate('-moz-opacity');
+      const hyphenatedStyleName = utils.hyphenate('-moz-opacity');
       assert.equal(hyphenatedStyleName, '-moz-opacity');
     });
   });
 
   describe('#createMarkup', () => {
     it('creates markup', () => {
-      const markup = createMarkup({opacity: 0});
+      const markup = utils.createMarkup({opacity: 0});
       assert.equal(markup, 'opacity:0;')
     });
 
     it('creates hyphenate css property', () => {
-      const markup = createMarkup({fontSize: '10px'});
+      const markup = utils.createMarkup({fontSize: '10px'});
       assert.equal(markup, 'font-size:10px;')
     });
 
     it('converts numbers to px values', () => {
-      const markup = createMarkup({fontSize: 10});
+      const markup = utils.createMarkup({fontSize: 10});
       assert.equal(markup, 'font-size:10px;');
     });
 
     it('creates no markup for empty object', () => {
-      const markup = createMarkup({});
+      const markup = utils.createMarkup({});
       assert.equal(markup, '')
     });
   });
