+++ lib/utils.js
@@ -1,4 +1,4 @@
-import { createMarkupForStyles } from 'react/lib/CSSPropertyOperations';
+import { isPropertyUnitfull } from './properties';
 
 export function sortObject(obj) {
   return Object.keys( obj )
@@ -54,9 +54,9 @@ export function createClassName(obj) {
     : undefined;
 }
 
-export function createMarkup(obj) {
-  return createMarkupForStyles( obj );
-}
+//export function createMarkup(obj) {
+//  return createMarkupForStyles( obj );
+//}
 
 export function isEmpty(obj) {
   return !Object.keys(obj).length;
@@ -106,3 +106,26 @@ export function seperateStyles (styles) {
     mediaQueries: []
   });
 }
+
+const isNumber = (value) => !isNaN(value);
+
+export const hyphenate = (stringToBeHyphenated) => {
+  return stringToBeHyphenated
+    .replace( /([a-z])([A-Z])/g, '$1-$2' )
+    .toLowerCase();
+};
+
+export const createMarkup = (styleDeclaration) => {
+  const rules = Object.keys(styleDeclaration).map((cssProperty) => {
+    const rawCssValue = styleDeclaration[cssProperty];
+    const hyphenatedProperty = hyphenate(cssProperty);
+
+    let cssValue = rawCssValue;
+    if(isNumber(cssValue) && isPropertyUnitfull(hyphenatedProperty)) {
+      cssValue += 'px';
+    }
+
+    return `${hyphenatedProperty}:${cssValue};`;
+  });
+  return rules.join('');
+};
